+++ src/com/mjr/games/DiceGame.java
@@ -6,17 +6,17 @@ import com.mjr.files.PointsSystem;
 public class DiceGame {
     public static final int DEFAULT_MULTIIPLER = 1;
 
-    public static int getWinPercent(int multi) {
+    public static double getWinPercent(double multi) {
 	return (100 - DEFAULT_MULTIIPLER) / multi;
     }
 
-    public static void procressTurn(String sender, int wager, int multi) {
+    public static void procressTurn(String sender, int wager, double multi) {
 	int randomNum = Utilities.getRandom(1, 100);
 	if (PointsSystem.hasPoints(sender, wager)) {
 	    PointsSystem.RemovePoints(sender, wager);
 
 	    if (randomNum > getWinPercent(multi)) {
-		int profit = wager * multi;
+		int profit = (int) (wager * multi);
 		PointsSystem.AddPoints(sender, profit);
 		try {
 		    Thread.sleep(5000);
